models: --resizebranch accepts floats; restore BEST filopodia run line
--resizebranch feeds nn.Upsample(scale_factor=(1, 1, r)) behind a `!= 1`
guard in every model that defines it, so fractional rates are meaningful
(0.5 halves the latent Z before net_g). argparse declared it type=int,
which rejected them outright. Widen to type=float in the 11 models that
still had int; MSclean already carried the fix locally.

run.sh: restore the MSclean/maxL1_b4Lb5 filopodia command, recovered from
its MLflow config.json artifact (the run's `bash` param is truncated at
250 chars). Its logged git_hash predates the float fix, so that commit
alone does not reproduce the line.

inference/README.md: train_mode must always be True in testing — with
--norm batch at -b 2 the BN running stats are too poorly conditioned to
reproduce training-time behaviour.

+++ models/ae0iso0tccut.py
@@ -118,7 +118,7 @@ class GAN(BaseModel):
         parser.add_argument("--c_mlp", dest='c_mlp', type=int, default=256, help='channel of mlp')
         parser.add_argument('--fWhich', nargs='+', help='which layers to have NCE loss', type=int, default=None)
         parser.add_argument("--downbranch", type=int, default=1)
-        parser.add_argument("--resizebranch", type=int, default=1)
+        parser.add_argument("--resizebranch", type=float, default=1.0)
         return parent_parser
 
     def encode(self, x):
